+++ src/types/date.d.ts
@@ -1,3 +1,4 @@
 import { Decoder } from './types';
 
 export const date: Decoder<Date>;
+export const iso8601: Decoder<Date>;