+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Upcoming
+
 ## v1.21.0
 
 -   **New decoder** [`lazy()`](https://github.com/nvie/decoders#lazy): lazily-evaluated